+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Adds `-v` and `--verbose` for printing out the version and path switched to.  
   By default, however, no news is good news.
 - Lists available Node versions when unable to locate the version requested.
+- Adds support for matching the latest version by only specifying the major or
+  minor version.  
+  E.g. to match `0.12.2` you can use `chnode 0` or `chnode 0.12`.
 
 ## 1.1.0 (Apr 22, 2015)
 - Detects current version correctly should it contain a revision number